LayoutTests/platform/chromium/http/tests/security/dataURL/xss-DENIED-to-data-url-from-data-url-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL data:text/html,<html><body><p id='accessMe'>PASS: Cross frame access from a data: URL was denied.</p><p>Inner iframe.</p><iframe name='innerFrame' id='innerFrame' src='data:text/html,<html><head><script>function test() {try {parent.document.getElementById("accessMe").innerHTML = "FAIL: Cross frame access from a data: URL was allowed.";} catch (e) {}if (window.layoutTestController)layoutTestController.notifyDone();}</script></head><body onload="test();"><p>Inner-inner iframe.</p></body></html>'></iframe></body></html> from frame with URL data:text/html,<html><head><script>function test() {try {parent.document.getElementById("accessMe").innerHTML = "FAIL: Cross frame access from a data: URL was allowed.";} catch (e) {}if (window.layoutTestController)layoutTestController.notifyDone();}</script></head><body onload="test();"><p>Inner-inner iframe.</p></body></html>. Domains, protocols and ports must match.

This tests that a data: URL loaded in an iframe inside another data: URL loaded iframe doesn't have access to its parent, the first data: URL loaded iframe.



--------
Frame: 'aFrame'
--------
PASS: Cross frame access from a data: URL was denied.

Inner iframe.



--------
Frame: 'innerFrame'
--------
Inner-inner iframe.
